check if file desc is up-to-date before uploading file

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/uploader/HTTPUploader.java
@@ -126,7 +126,7 @@ public class HTTPUploader extends AbstractUploader implements Uploader {
                 setUploadBegin(request.getLow());
                 setUploadEnd(request.getHigh() + 1);
             } else {
-                if (!ifd.isRangeSatisfiable((int) first, (int) last - 1)) {
+                if (!ifd.isRangeSatisfiable(first, last - 1)) {
                     return false;
                 } 
             }
